fix data type number 0f floor in feature

+++ Modules/Appraisal/Appraisal/Application/Features/Appraisals/CreateLeaseAgreementLandAndBuildingProperty/CreateLeaseAgreementLandAndBuildingPropertyRequest.cs
@@ -128,7 +128,7 @@ public record CreateLeaseAgreementLandAndBuildingPropertyRequest(
     string? BuildingAreaUnit = null,
     decimal? UsableArea = null,
     // Building - Structure
-    int? NumberOfFloors = null,
+    decimal? NumberOfFloors = null,
     // Building - Style
     string? BuildingMaterialType = null,
     string? BuildingStyleType = null,
